REFACTOR: created equations.py module and mmoved relevant equations into it
